Warning fix
git-svn-id: https://svn.wxwidgets.org/svn/wx/wxWidgets/trunk@49719 c3d73ce0-8a6f-49c7-b76d-6d57e0e08775

+++ src/generic/choicbkg.cpp
@@ -109,7 +109,7 @@ wxChoicebook::Create(wxWindow *parent,
 
     m_controlSizer = new wxBoxSizer(IsVertical() ? wxHORIZONTAL : wxVERTICAL);
     m_controlSizer->Add(m_bookctrl, 1, (IsVertical() ? wxALIGN_CENTRE_VERTICAL : wxALIGN_CENTRE) |wxGROW, 0);
-    mainSizer->Add(m_controlSizer, 0, (IsVertical() ? wxGROW : wxALIGN_CENTRE_VERTICAL)|wxALL, m_controlMargin);
+    mainSizer->Add(m_controlSizer, 0, (IsVertical() ? (int) wxGROW : (int) wxALIGN_CENTRE_VERTICAL)|wxALL, m_controlMargin);
     SetSizer(mainSizer);
     return true;
 }
